tests: moved shared test fixtures to conftest.py
