+++ train_lib/tests/test_security_functions.py
@@ -37,7 +37,23 @@ def test_encryption_decryption(config_dict_as_json):
 
     assert not file_content_changed
 
-    # TODO test file encryption with temp files
+
+def test_encrypt_decrypt_train_files(train_files, symmetric_key, query_json):
+
+    fe = FileEncryptor(symmetric_key)
+    filenames, files = train_files
+    encrypted_files = fe.encrypt_files(files, binary_files=True)
+    encrypted_query = fe.encrypt_file(query_json)
+    encrypted_files.append(encrypted_query)
+
+    decrypted_files = fe.decrypt_files(encrypted_files, binary_files=True)
+    decrypted_files.append(fe.decrypt_file(encrypted_query))
+
+    for i, file in enumerate(files):
+        file.seek(0)
+        file_content = file.read()
+        decryted_content = decrypted_files[i].read()
+        assert decryted_content == file_content
 
 
 def test_aes_encryption():
